requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn>=0.21.1,<0.22.0
python-dotenv>=1.0.0,<2.0.0
langchain>=0.0.21,<0.1.0
langgraph>=0.150.16.0
azure-openai>=1.0.0,<2.0.0